requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
fastapi==0.109.0
fonttools==4.48.1
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
idna==3.4
kiwisolver==1.4.5
magic-filter==1.0.12
matplotlib==3.8.2
mplfinance==0.12.10b0
multidict==6.0.4
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.35.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.26.0
yarl==1.9.2
